boss_agent/requirements.txt
openai==0.27.0
websockets==10.4
python-dotenv==0.19.2
